requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.1.0
pyjwt[cryptography]==2.8.0
redis==5.0.1
prometheus-fastapi-instrumentator==6.0.0
python-multipart==0.0.6